connector/bdw-cp-connector/requirements.txt
apscheduler==3.10.4
Werkzeug==2.2.2
sqlalchemy-redshift==0.8.14
sqlalchemy-bigquery==1.8.0
snowflake-sqlalchemy==1.5.0
redshift-connector==2.0.913
snowflake-connector-python==3.1.1
#PyGreSQL==5.2.4
confluent-kafka==2.2.0
psycopg2-binary==2.9.5
google-cloud-bigquery-storage==2.22.0
pyarrow==13.0.0
pandas==2.1.0

# conda install
# pyodbc==4.0.39 
# pymssql==2.2.5

pyspark==3.4.1

Flask==2.1.2
Flask[async]
